manage user's access, timeline vue

+++ back-end/Users/mutation.py
@@ -24,11 +24,21 @@ class CreateUser(graphene.Mutation):
 
     user = graphene.Field(UserType)
     def mutate(self, info, username, password, first_name, last_name, phone, email, address , role):
-        user = Users(username=username, password=password, first_name=first_name, last_name=last_name, phone=phone , email=email, address=address, role=role)
-        user.set_password(password)
-        user.save()
-        send_create(user)
-        return CreateUser(user=user)
+        current_user = info.context.user
+        if not current_user.is_authenticated:
+            raise Exception("You should login")
+        
+        # if not current_user.is_staff:
+        #     raise Exception("You do not have access")
+        
+        if not current_user.role !="admin":
+            raise Exception("You do not have access")
+        
+        target_user = Users(username=username, password=password, first_name=first_name, last_name=last_name, phone=phone , email=email, address=address, role=role)
+        target_user.set_password(password)
+        target_user.save()
+        send_create(current_user,target_user)
+        return CreateUser(user=target_user)
     
 
 """
@@ -50,48 +60,52 @@ class UpdateUser(graphene.Mutation):
     user = graphene.Field(UserType)
     def mutate(self, info , id, username=None, password=None, first_name=None, last_name=None, phone=None, email=None, 
     address=None , role=None):
+        current_user = info.context.user
+        if not current_user.is_authenticated:
+            raise Exception("You should login")
+        
         changes = {}
         try:
-            user = Users.objects.get(pk=id)
-            if username is not None and user.username!=username:
-                changes["username"] = {"old_val": user.username , "new_val": username}
-                user.username = username
+            target_user = Users.objects.get(pk=id)
+            if username is not None and target_user.username!=username:
+                changes["username"] = {"old_val": target_user.username , "new_val": username}
+                target_user.username = username
 
-            if password is not None and not user.check_password(password):
+            if password is not None and not target_user.check_password(password):
                 changes["password"] = {"old_val": "********", "new_val": "********"}
-                user.set_password(password)
+                target_user.set_password(password)
 
 
-            if first_name is not None and user.first_name!=first_name:
-                changes["first_name"] = {"old_val": user.first_name , "new_val": first_name}
-                user.first_name = first_name
+            if first_name is not None and target_user.first_name!=first_name:
+                changes["first_name"] = {"old_val": target_user.first_name , "new_val": first_name}
+                target_user.first_name = first_name
 
-            if last_name is not None and user.last_name!=last_name:
-                changes["last_name"] = {"old_val": user.last_name , "new_val": last_name}
-                user.last_name = last_name
+            if last_name is not None and target_user.last_name!=last_name:
+                changes["last_name"] = {"old_val": target_user.last_name , "new_val": last_name}
+                target_user.last_name = last_name
 
-            if phone is not None and user.phone!=phone:
-                changes["phone"] = {"old_val": user.phone , "new_val": phone}
-                user.phone = phone
+            if phone is not None and target_user.phone!=phone:
+                changes["phone"] = {"old_val": target_user.phone , "new_val": phone}
+                target_user.phone = phone
 
-            if email is not None and user.email!=email:
-                changes["email"] = {"old_val": user.email , "new_val": email}
-                user.email = email
+            if email is not None and target_user.email!=email:
+                changes["email"] = {"old_val": target_user.email , "new_val": email}
+                target_user.email = email
 
-            if address is not None and user.address!=address:
-                changes["address"] = {"old_val": user.address , "new_val": address}
-                user.address = address
+            if address is not None and target_user.address!=address:
+                changes["address"] = {"old_val": target_user.address , "new_val": address}
+                target_user.address = address
             
-            if role is not None and user.role!=role:
-                changes["role"] = {"old_val": user.role , "new_val": role}
-                user.role = role
+            if role is not None and target_user.role!=role:
+                changes["role"] = {"old_val": target_user.role , "new_val": role}
+                target_user.role = role
 
-            user.save()
-            send_edit(user,"SUCCESS", changes=changes)
-            return UpdateUser(user=user)
+            target_user.save()
+            send_edit(current_user,target_user,"SUCCESS", changes=changes)
+            return UpdateUser(user=target_user)
     
         except Exception as e:
-            send_edit(user, "FAILED" ,changes=changes)
+            send_edit(current_user,target_user, "FAILED" ,changes=changes)
             raise Exception(f"Update User Failed:{str(e)}")
 
 """
@@ -101,10 +115,30 @@ class DeleteUser(graphene.Mutation):
     class Arguments:
         id = graphene.ID(required=True)
     ok = graphene.Boolean()
+
     def mutate(self , info, id):
-        user = Users.objects.get(pk=id)
-        user.delete()
-        send_delete(user)
+        current_user = info.context.user
+        if not current_user.is_authenticated:
+            raise Exception("You should login")
+        
+        try:
+            target_user = Users.objects.get(pk=id)
+        except:
+            raise Exception("target user not found")
+        
+        if current_user.id==target_user.id:
+            target_user.delete()
+            send_delete(target_user)
+            return DeleteUser(ok=True)
+        
+        if current_user.role!="admin":
+            raise Exception("You do noy have access")
+        
+        if target_user.role=="admin":
+            raise Exception("You con not delete admin users")
+        
+        target_user.delete()
+        send_delete(current_user, target_user)
         return DeleteUser(ok=True)
 
 class LoginUser(graphene.Mutation):
@@ -154,31 +188,34 @@ class UserMutation(graphene.ObjectType):
 
 
 
-def send_edit(user , action_s , changes):
-    description = f"changes profile of user: {user}\n"
+def send_edit(actoruser , targetuser, action_s , changes):
+    description = f"changes profile of user: {targetuser}\n"
     for key,val in changes.items():
         description += f"_{key}: old value {val['old_val']} changes to new value{val['new_val']} \n"
        
         
     Logs.objects.create(
-        user=user,
+        target_user=targetuser,
+        actor_user=actoruser,
         action_type="UPDATE_PROFILE",
         action_status=action_s,
         description=description,
     )
 
-def send_delete(user):
+def send_delete(actoruser, targetuser):
     Logs.objects.create(
-        user=user,
+        actor_user=actoruser,
+        target_user=targetuser,
         action_type="DELETE",
-        description=f"delete {user} ",
+        description=f"delete {targetuser} ",
     )
         
-def send_create(user):
+def send_create(actoruser, targetuser ):
     Logs.objects.create(
-        user=user,
+        actor_user=actoruser,
+        target_user=targetuser,
         action_type="CREATE",
-        description=f"create {user}",
+        description=f"create {targetuser}",
     )
 
 def send_login(user):
